pyproject.toml
[tool.poetry]
name = "crypto-lob-micro-move"
version = "0.1.0"
description = "Crypto LOB micro-move classifier (BTC/ETH/ADA)"
authors = ["Ahmet <ahmet@example.com>"]

packages = [{ include = "crypto_lob", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "2.2.2"
pytorch-lightning = "^2.2"
torchmetrics = "^1.3"
pandas = "^2.2"
numpy = "^1.26"
hydra-core = "^1.3"
omegaconf = "^2.3"
mlflow = "^2.12"
dvc = "^3.42"
fire = "^0.6"
scikit-learn = "^1.4"
rich = "^13.7"
kagglehub = { git = "https://github.com/Kaggle/kagglehub.git", branch = "main" }

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6"
pytest = "^8.0"
black = "^24.3"
isort = "^5.13"
flake8 = "^7.0"

[tool.poetry.scripts]
clob = "crypto_lob.commands:entry"
